Moved some function defs to their own cpp file
