+++ src/axom/klee/io/IOUtil.cpp
@@ -9,29 +9,10 @@
 #include "axom/inlet.hpp"
 #include "axom/klee/KleeError.hpp"
 
-#include <stdexcept>
-
 namespace axom
 {
 namespace klee
 {
-LengthUnit parseLengthUnits(const std::string &unitsAsString, const std::string &path)
-{
-  try
-  {
-    return utilities::getLengthUnit(unitsAsString);
-  }
-  catch(const std::invalid_argument &ex)
-  {
-    throw KleeError({path, ex.what()});
-  }
-}
-
-LengthUnit parseLengthUnits(const inlet::Proxy &unitsAsProxy)
-{
-  return parseLengthUnits(unitsAsProxy.get<std::string>(), unitsAsProxy.name());
-}
-
 namespace internal
 {
 std::vector<double> toDoubleVector(inlet::Proxy const &field,
